requirements.txt
PyMySQL==0.10.0
Flask==1.1.2
cx_Oracle==8.1.0
python_Levenshtein==0.12.2
